Improve Week 2: Fibonacci mod m
The previous code built an array for all values in the pisano period.
In theory, this array could be used to directly lookup the result.
In practice, it's memory that needs to be kept around (with m-square
elements at most).
This commit removes the array and directly returns the Pisano
sequence number (i.e. the length of the sequence) itself.