Add Week 4: Organizing a lottery
This one is an implementation of a buddy's suggestion (Kudos @baracoder)
to first reduce the start and end times into a list of delta values,
accumulating them to get a list of total overlaps per time and then
just looking up the query points in the accumulated list.
While the first steps simplifies the list representation, the last
step can be done efficiently using a binary search.
In this implementation, the framework's BinarySearch list method is
used with a custom comparer.

One thing to be aware of here is that the end time is still considered
to be part of the line segment, so decreasing the accumulated count
can only happen AFTER the end time. Since all times are discrete,
this is easily solved by decreasing the accumulated count at exactly
one time step later.